add UnitInitEtc

+++ Src/Object/UnitBase.h
@@ -96,6 +96,6 @@ protected:
 	std::unordered_map<ANIM, int> animNum_;	//アニメーションナンバー格納配列。
 	int atcAnim_;			//アタッチするアニメを格納
 	int animTotalTime_;		//アニメーションの総再生時間
-	int stepAnim_;			//アニメーションの再生時間
+	float stepAnim_;			//アニメーションの再生時間
 };
